# Prevent integer width conversion in caml_int_compare