--- a/runtime/floats.c
+++ b/runtime/floats.c
@@ -1008,8 +1008,15 @@ intnat caml_float_compare_unboxed(double f, double g)
   /* If one or both of f and g is NaN, order according to the convention
      NaN = NaN and NaN < x for all other floats x. */
   /* This branchless implementation is from GPR#164.
-     Note that [f == f] if and only if f is not NaN. */
-  return (f > g) - (f < g) + (f == f) - (g == g);
+     Note that [f == f] if and only if f is not NaN.
+     We expand each subresult of the expression to
+     avoid sign-extension on 64bit. GPR#2250. */
+  intnat greater = f > g;
+  intnat less    = f < g;
+  intnat equal_f = f == f;
+  intnat equal_g = g == g;
+  intnat res = greater - less + equal_f - equal_g;
+  return res;
 }
 
 #endif