Previously, in 64bit mode, res was computed as 32bits, and then extended to 64bits.
This patch eliminates the sign-extension instruction.