@gasche 2019-02-15
Could you include a comment in the code to record the intent in this change?