@xclerc 2019-02-15
This code pattern has a number of occurrences,
*e.g.* caml_int32_compare_unboxed_again.
Shouldn't we update all of them? (Possibly by
defining a macro.)