@xavierleroy 2019-02-15
Does this change improves performance in a measurable way?  int32 -> int64 conversions are extremely cheap, and the proposed code is less readable than the current code...  I'm not asking for big speedups, just a hint that it makes a non-negligible difference.