@smuenzel-js 2019-02-18
I've updated the code to use a macro, and changed all the other comparisons that had this issue.

Benchmark results on a loop that adds up the results of comparisons:
```
New: 
./a.out  15.00s user 0.00s system 99% cpu 15.011 total
./a.out  14.86s user 0.01s system 99% cpu 14.874 total
./a.out  14.94s user 0.00s system 99% cpu 14.947 total

Old:
./a.out  17.76s user 0.01s system 99% cpu 17.767 total
./a.out  17.30s user 0.00s system 99% cpu 17.316 total
./a.out  18.09s user 0.01s system 99% cpu 18.100 total
```

This is the program:
```
let do_compare (x : int) (y : int) =
  let x = Sys.opaque_identity x in
  let y = Sys.opaque_identity y in
  compare x y

let () =
  let result = ref 0 in
  for i = 0 to 4_000_000_000 do
    result := !result + (do_compare 432 5);
    result := !result + (do_compare 5 432);
    result := !result + (do_compare 432 432);
  done;
  print_int !result
```

The difference in the two functions is clear when you look at the disassembly.
Previously:
```
XDIS 0000000000418630: BASE 31C0                     xor eax, eax
XDIS 0000000000418632: BASE 4839F7                   cmp rdi, rsi
XDIS 0000000000418635: BASE 0F9FC0                   setnle al
XDIS 0000000000418638: BASE 31D2                     xor edx, edx
XDIS 000000000041863a: BASE 4839F7                   cmp rdi, rsi
XDIS 000000000041863d: BASE 0F9CC2                   setl dl
XDIS 0000000000418640: BASE 29D0                     sub eax, edx
XDIS 0000000000418642: LONGMODE 4898                     cdqe 
XDIS 0000000000418644: BASE 488D440001               lea rax, ptr [rax+rax*1+0x1]
XDIS 0000000000418649: BASE C3                       ret 
```
With this patch:
```
XDIS 0000000000418430: BASE 31C0                     xor eax, eax
XDIS 0000000000418432: BASE 4839F7                   cmp rdi, rsi
XDIS 0000000000418435: BASE 0F9CC2                   setl dl
XDIS 0000000000418438: BASE 0F9FC0                   setnle al
XDIS 000000000041843b: BASE 0FB6D2                   movzx edx, dl
XDIS 000000000041843e: BASE 4829D0                   sub rax, rdx
XDIS 0000000000418441: BASE 488D440001               lea rax, ptr [rax+rax*1+0x1]
XDIS 0000000000418446: BASE C3                       ret 
```