@xclerc 2019-02-18
Thanks for the changes; Is there a reason to not apply it
to `{int32,int64,nativeint}_cmp`?