@xclerc 2019-02-18
Answering my own question: the aforementioned functions
do not need to be modified because they return an `int`.