@xavierleroy 2019-02-22
I liked it better before the last commit.  Could we please stop suggesting things and pushing code, and instead proceed with a proper review?  Which means that the code should not change every day.