@smuenzel-js 2019-03-04
@xavierleroy I've reverted the latest change.